# Use SHA2 and PSS for client assertion