--- a/src/client/Microsoft.Identity.Client/AppConfig/AuthorityInfo.cs
+++ b/src/client/Microsoft.Identity.Client/AppConfig/AuthorityInfo.cs
@@ -311,8 +311,6 @@ namespace Microsoft.Identity.Client
             }
         }
 
-        
-
         internal Authority CreateAuthority()
         {
             switch (AuthorityType)