--- a/src/client/Microsoft.Identity.Client/Internal/ClientCredential/SignedAssertionClientCredential.cs
+++ b/src/client/Microsoft.Identity.Client/Internal/ClientCredential/SignedAssertionClientCredential.cs
@@ -22,7 +22,15 @@ namespace Microsoft.Identity.Client.Internal.ClientCredential
             _signedAssertion = signedAssertion;
         }
 
-        public Task AddConfidentialClientParametersAsync(OAuth2Client oAuth2Client, ILoggerAdapter logger, ICryptographyManager cryptographyManager, string clientId, string tokenEndpoint, bool sendX5C, CancellationToken cancellationToken)
+        public Task AddConfidentialClientParametersAsync(
+            OAuth2Client oAuth2Client, 
+            ILoggerAdapter logger, 
+            ICryptographyManager cryptographyManager, 
+            string clientId, 
+            string tokenEndpoint,
+            bool sendX5C,
+            bool useSha2,
+            CancellationToken cancellationToken)
         {
             oAuth2Client.AddBodyParameter(OAuth2Parameter.ClientAssertionType, OAuth2AssertionType.JwtBearer);
             oAuth2Client.AddBodyParameter(OAuth2Parameter.ClientAssertion, _signedAssertion);