--- a/src/client/Microsoft.Identity.Client/Internal/ClientCredential/SignedAssertionDelegateClientCredential.cs
+++ b/src/client/Microsoft.Identity.Client/Internal/ClientCredential/SignedAssertionDelegateClientCredential.cs
@@ -28,7 +28,15 @@ namespace Microsoft.Identity.Client.Internal.ClientCredential
             _signedAssertionWithInfoDelegate = signedAssertionDelegate;
         }
 
-        public async Task AddConfidentialClientParametersAsync(OAuth2Client oAuth2Client, ILoggerAdapter logger, ICryptographyManager cryptographyManager, string clientId, string tokenEndpoint, bool sendX5C, CancellationToken cancellationToken)
+        public async Task AddConfidentialClientParametersAsync(
+            OAuth2Client oAuth2Client, 
+            ILoggerAdapter logger, 
+            ICryptographyManager cryptographyManager,
+            string clientId, 
+            string tokenEndpoint,
+            bool sendX5C, 
+            bool useSha2,
+            CancellationToken cancellationToken)
         {
             string signedAssertion = await (_signedAssertionDelegate != null 
                 ? _signedAssertionDelegate(cancellationToken).ConfigureAwait(false)
@@ -41,5 +49,6 @@ namespace Microsoft.Identity.Client.Internal.ClientCredential
             oAuth2Client.AddBodyParameter(OAuth2Parameter.ClientAssertionType, OAuth2AssertionType.JwtBearer);
             oAuth2Client.AddBodyParameter(OAuth2Parameter.ClientAssertion, signedAssertion);
         }
+
     }
 }