--- a/src/client/Microsoft.Identity.Client/Platforms/net6/MsalJsonSerializerContext.cs
+++ b/src/client/Microsoft.Identity.Client/Platforms/net6/MsalJsonSerializerContext.cs
@@ -35,7 +35,6 @@ namespace Microsoft.Identity.Client.Platforms.net6
     [JsonSerializable(typeof(UserRealmDiscoveryResponse))]
     [JsonSerializable(typeof(DeviceCodeResponse))]
     [JsonSerializable(typeof(AdfsWebFingerResponse))]
-    [JsonSerializable(typeof(JsonWebToken.JWTPayload))]
     [JsonSerializable(typeof(DeviceAuthHeader))]
     [JsonSerializable(typeof(DeviceAuthPayload))]
     [JsonSerializable(typeof(ManagedIdentityResponse))]