… src/client/Microsoft.Identity.Client/OAuth2/TokenClient.cs
… src/client/Microsoft.Identity.Client/Platforms/uap/UapCryptographyManager.cs
… src/client/Microsoft.Identity.Client/PlatformsCommon/Interfaces/ICryptographyManager.cs
… src/client/Microsoft.Identity.Client/PlatformsCommon/Shared/CommonCryptographyManager.cs
… src/client/Microsoft.Identity.Client/PlatformsCommon/Shared/DeviceAuthManager.cs
… src/client/Microsoft.Identity.Client/Utils/DateTimeHelpers.cs
… tests/Microsoft.Identity.Test.Common/TestConstants.cs
… tests/Microsoft.Identity.Test.Core.UIAutomation/Microsoft.Identity.Test.UIAutomation.Infrastructure.csproj
… tests/Microsoft.Identity.Test.Integration.netfx/HeadlessTests/ClientCredentialsTests.WithRegion.cs
… tests/Microsoft.Identity.Test.Integration.netfx/HeadlessTests/ClientCredentialsTests.cs
… tests/Microsoft.Identity.Test.Integration.netfx/SeleniumTests/ConfidentialClientAuthorizationTests.cs
… tests/Microsoft.Identity.Test.Performance/ClientAssertionTests.cs
… tests/Microsoft.Identity.Test.Unit/ApiConfigTests/AuthorityTests.cs
… tests/Microsoft.Identity.Test.Unit/CryptographyTests.cs
… tests/Microsoft.Identity.Test.Unit/PublicApiTests/ClientCredentialWithCertTest.cs
… tests/Microsoft.Identity.Test.Unit/TestBase.cs
… tests/Microsoft.Identity.Test.Unit/UtilTests/JsonHelperTests.cs